requirements.txt
openai
streamlit
python-dotenv
redis
scipy
pydantic
newsapi-python
python-docx
asyncio
crawl4ai
beautifulsoup4
Markdown
requests
mcp-use